mod net architecture for vlp 32

+++ src/eval.py
@@ -19,54 +19,47 @@ from config import *
 from imdb import kitti
 from utils.util import *
 from nets import *
-# from sklearn.metrics import confusion_matrix
+from sklearn.metrics import confusion_matrix
 
 FLAGS = tf.app.flags.FLAGS
 
-tf.app.flags.DEFINE_string('dataset', 'KITTI',
-                           """Currently support KITTI dataset.""")
+tf.app.flags.DEFINE_string('dataset', 'KITTI', """Currently support KITTI dataset.""")
 tf.app.flags.DEFINE_string('data_path', '', """Root directory of data""")
-tf.app.flags.DEFINE_string('image_set', 'val',
-                           """Can be train, trainval, val, or test""")
+tf.app.flags.DEFINE_string('image_set', 'val', """Can be train, trainval, val, or test""")
 tf.app.flags.DEFINE_string('eval_dir', '/tmp/bichen/logs/squeezeSeg/eval',
                             """Directory where to write event logs """)
-tf.app.flags.DEFINE_string('checkpoint_path', '/tmp/bichen/logs/squeezeSeg/train',
-                            """Path to the training checkpoint.""")
-tf.app.flags.DEFINE_integer('eval_interval_secs', 60 * 1,
-                             """How often to check if new cpt is saved.""")
-tf.app.flags.DEFINE_boolean('run_once', False,
-                             """Whether to run eval only once.""")
-tf.app.flags.DEFINE_string('net', 'squeezeSeg',
-                           """Neural net architecture.""")
+tf.app.flags.DEFINE_string('checkpoint_path', '',"""Path to the training checkpoint.""")
+tf.app.flags.DEFINE_integer('eval_interval_secs', 60 * 1, """How often to check if new cpt is saved.""")
+tf.app.flags.DEFINE_boolean('run_once', False, """Whether to run eval only once.""")
+tf.app.flags.DEFINE_string('net', 'squeezeSeg', """Neural net architecture.""")
 tf.app.flags.DEFINE_string('gpu', '0', """gpu id.""")
+tf.app.flags.DEFINE_string('extended', 'y', """Extended classes.""")
+tf.app.flags.DEFINE_string('restore', 'n', """Start from checkpoint""")
 
-# def print_cm(cm, labels, hide_zeroes=False, hide_diagonal=False, hide_threshold=None):
-#     """pretty print for confusion matrixes"""
-#     columnwidth = max([len(x) for x in labels] + [5])  # 5 is value length
-#     empty_cell = " " * columnwidth
-#     # Print header
-#     print("    " + empty_cell, end=" ")
-#     for label in labels:
-#         print("%{0}s".format(columnwidth) % label, end=" ")
-#     print()
-#     # Print rows
-#     for i, label1 in enumerate(labels):
-#         print("    %{0}s".format(columnwidth) % label1, end=" ")
-#         for j in range(len(labels)):
-#             cell = "%{0}.4f".format(columnwidth) % cm[i, j]
-#             if hide_zeroes:
-#                 cell = cell if float(cm[i, j]) != 0 else empty_cell
-#             if hide_diagonal:
-#                 cell = cell if i != j else empty_cell
-#             if hide_threshold:
-#                 cell = cell if cm[i, j] > hide_threshold else empty_cell
-#             print(cell, end=" ")
-#         print()
+def print_cm(cm, labels, hide_zeroes=False, hide_diagonal=False, hide_threshold=None):
+    """pretty print for confusion matrixes"""
+    columnwidth = max([len(x) for x in labels] + [5])  # 5 is value length
+    empty_cell = " " * columnwidth
+    # Print header
+    print("    " + empty_cell, end=" ")
+    for label in labels:
+        print("%{0}s".format(columnwidth) % label, end=" ")
+    print()
+    # Print rows
+    for i, label1 in enumerate(labels):
+        print("    %{0}s".format(columnwidth) % label1, end=" ")
+        for j in range(len(labels)):
+            cell = "%{0}.4f".format(columnwidth) % cm[i, j]
+            if hide_zeroes:
+                cell = cell if float(cm[i, j]) != 0 else empty_cell
+            if hide_diagonal:
+                cell = cell if i != j else empty_cell
+            if hide_threshold:
+                cell = cell if cm[i, j] > hide_threshold else empty_cell
+            print(cell, end=" ")
+        print()
 
-def eval_once(
-    saver, ckpt_path, summary_writer, eval_summary_ops, eval_summary_phs, imdb,
-    model):
-
+def eval_once(saver, ckpt_path, summary_writer, eval_summary_ops, eval_summary_phs, imdb, model):
   with tf.Session(config=tf.ConfigProto(allow_soft_placement=True)) as sess:
 
     # Restores from checkpoint
@@ -102,6 +95,11 @@ def eval_once(
     ofn_sum = np.zeros(mc.NUM_CLASS)
     ofp_sum = np.zeros(mc.NUM_CLASS)
 
+    if mc.EVAL_ON_ORG:
+      cm = np.zeros([4, 4])
+    else: 
+      cm = np.zeros([mc.NUM_CLASS, mc.NUM_CLASS])
+
     for i in xrange(int(num_images/mc.BATCH_SIZE)):
       offset = max((i+1)*mc.BATCH_SIZE - num_images, 0)
       
@@ -122,6 +120,17 @@ def eval_once(
       _t['detect'].toc()
 
       _t['eval'].tic()
+
+      if  mc.EVAL_ON_ORG:
+        print('Translating labels...')
+        pred_cls[(pred_cls!=0)*(pred_cls!=7)*(pred_cls!=6)*(pred_cls!=9)] =0 
+        pred_cls[pred_cls==6]=2
+        pred_cls[pred_cls==9]=3
+        pred_cls[pred_cls==7]=1           
+        cm+=confusion_matrix(y_true=label_per_batch.flatten(), y_pred=pred_cls.flatten(), labels=range(4))
+      else:
+        cm+=confusion_matrix(y_true=label_per_batch.flatten(), y_pred=pred_cls.flatten(), labels=range(mc.NUM_CLASS))
+      
       # Evaluation
       iou, tps, fps, fns = evaluate_iou(
           label_per_batch[:mc.BATCH_SIZE-offset],
@@ -161,7 +170,7 @@ def eval_once(
 
     if mc.EVAL_ON_ORG:
         print ('  Accuracy:')
-        classes=['unknown' ,  'car', 'pedestrian'  ,  'cyclist']
+        classes=['unknown' ,  'car', 'pedestrian'  ,  'cyclist'] 
         for i in range(1, 4):
             print ('    {}:'.format(classes[i]))
             print ('\tPixel-seg: P: {:.3f}, R: {:.3f}, IoU: {:.3f}'.format(
@@ -182,32 +191,47 @@ def eval_once(
     eval_summary_str = sess.run(eval_summary_ops, feed_dict=eval_sum_feed_dict)
     for sum_str in eval_summary_str:
       summary_writer.add_summary(sum_str, global_step)
-    # summary_writer.flush()
-    # cm = cm / cm.sum(axis=1)[:, np.newaxis] 
-    # if  mc.EVAL_ON_ORG:
-    #     print_cm(cm, ['unknown' ,  'car', 'pedestrian'  ,  'cyclist'])
-    # else:
-    #     print_cm(cm, mc.CLASSES[i])
+    summary_writer.flush()
+    cm = cm / cm.sum(axis=1)[:, np.newaxis] 
+    if  mc.EVAL_ON_ORG:
+        print_cm(cm, ['unknown' ,  'car', 'pedestrian'  ,  'cyclist'])
+    else:
+        print_cm(cm, mc.CLASSES[i])
 
 def evaluate():
   """Evaluate."""
   assert FLAGS.dataset == 'KITTI', \
       'Currently only supports KITTI dataset'
 
-  # os.environ['CUDA_VISIBLE_DEVICES'] = FLAGS.gpu
-  os.environ['CUDA_VISIBLE_DEVICES'] = ""
+  if FLAGS.gpu == '0':
+    os.environ['CUDA_VISIBLE_DEVICES'] = FLAGS.gpu
+  else:
+    os.environ['CUDA_VISIBLE_DEVICES'] = "" # Train only with CPU
 
   with tf.Graph().as_default() as g:
 
-    assert FLAGS.net == 'squeezeSeg', \
+    assert FLAGS.net == 'squeezeSeg' or FLAGS.net == 'squeezeSeg32', \
         'Selected neural net architecture not supported: {}'.format(FLAGS.net)
 
-    if FLAGS.net == 'squeezeSeg':
-      # mc = kitti_squeezeSeg_config()
-      mc = kitti_squeezeSeg_config_ext()
-      mc.LOAD_PRETRAINED_MODEL = False
+    if FLAGS.net == 'squeezeSeg':    
+      if FLAGS.extended == 'y':
+        mc = kitti_squeezeSeg_config_ext() # Added ground class
+      else:
+        mc = kitti_squeezeSeg_config()   # Original training set
+
+      mc.PRETRAINED_MODEL_PATH = False #FLAGS.pretrained_model_path
       mc.BATCH_SIZE = 1 # TODO(bichen): fix this hard-coded batch size.
       model = SqueezeSeg(mc)
+
+    elif FLAGS.net == 'squeezeSeg32':    
+      if FLAGS.extended == 'y':
+        mc = kitti_squeezeSeg32_config_ext() # Added ground class
+      else:
+        mc = kitti_squeezeSeg32_config()   # Original training set
+
+      mc.PRETRAINED_MODEL_PATH = False #FLAGS.pretrained_model_path
+      mc.BATCH_SIZE = 1 # TODO(bichen): fix this hard-coded batch size.
+      model = SqueezeSeg32(mc)
 
     imdb = kitti(FLAGS.image_set, FLAGS.data_path, mc)
 
@@ -266,9 +290,9 @@ def evaluate():
 
 
 def main(argv=None):  # pylint: disable=unused-argument
-  if tf.gfile.Exists(FLAGS.eval_dir):
+  if tf.gfile.Exists(FLAGS.eval_dir) and FLAGS.restore == 'n':
     tf.gfile.DeleteRecursively(FLAGS.eval_dir)
-  tf.gfile.MakeDirs(FLAGS.eval_dir)
+    tf.gfile.MakeDirs(FLAGS.eval_dir)
   evaluate()
